requirements.txt
google-generativeai>=0.7.0,<0.8.0
chromadb==0.5.3
pdf2image==1.16.3
pypdf==4.3.1
python-dotenv==1.0.1
langchain==0.3.0
langchain-google-genai==2.0.0
streamlit==1.39.0
sentence-transformers==3.0.0